+++ nuget-client/BrevgeneratorClient/IBrevgeneratorClient.cs
@@ -1,8 +1,8 @@
-﻿using AT.Brevgenerator.Klient.Model;
+﻿using Arbeidstilsynet.Brevgenerator.Client.Models;
 
-namespace AT.Brevgenerator.Klient
+namespace Arbeidstilsynet.Brevgenerator.Client
 {
-    public interface IBrevgeneratorKlient
+    public interface IBrevgeneratorClient
     {
         /// <summary>
         /// Generer et brev ut i fra dynamisk markdown, variabler (flettefelt) og diverse konfigurasjon.<br />