+++ nuget-client/BrevgeneratorClient/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Versjonslogg
 
+## 4.0.0
+
+- Publiserer nå til `nuget.org` i stedet for Azure DevOps
+- Omdøpt pakke fra `AT.Brevgenerator.Klient` til `Arbeidstilsynet.Brevgenerator.Client`
+- Omdøpt namespace fra `AT.Brevgenerator.Klient` / `AT.Brevgenerator.Klient.Model` til `Arbeidstilsynet.Brevgenerator.Client` / `Arbeidstilsynet.Brevgenerator.Client.Models`
+- Omdøpt `BrevgeneratorKlient` / `IBrevgeneratorKlient` til `BrevgeneratorClient` / `IBrevgeneratorClient`
+
 ## 3.3.0
 
 - Lagt til støtte for `ErUnntattOffentlighet` og `UnntattOffentlighetHjemmel` i direktorat-template